Add authorization on video services

+++ apps/rest/src/main/scala/by/sideproject/videocaster/app/rest/routes/video/handlers/VideosDeleteEntityHandler.scala
@@ -3,11 +3,10 @@ package by.sideproject.videocaster.app.rest.routes.video.handlers
 import akka.actor.Actor
 import by.sideproject.instavideo.filestorage.base.FileStorageService
 import by.sideproject.videocaster.app.rest.routes.base.requests.EntityRequest
-import by.sideproject.videocaster.app.rest.routes.video.requests.VideosGetEntitiesRequest
-import by.sideproject.videocaster.model.VideoItemDetails
 import by.sideproject.videocaster.services.storage.base.dao.VideoItemDetailsDAO
 import org.slf4j.LoggerFactory
 import spray.http.StatusCodes
+import spray.routing.AuthorizationFailedRejection
 
 import scala.concurrent.ExecutionContext
 
@@ -25,12 +24,18 @@ class VideosDeleteEntityHandler(videoDetailsDAO: VideoItemDetailsDAO, binaryStor
       for {
         videoItemForRemoval <- videoDetailsDAO.findOneById(id)
       } yield {
-        videoItemForRemoval.flatMap {
-          videoDetailsDAO.removeById(id)
-          _.fileMetaId
-        }.map(meta => binaryStorageService.remove(meta, identity))
-
-        ctx.complete(videoItemForRemoval)
+        videoItemForRemoval match {
+          case Some(item) => {
+            if (item.profileId == identity.profileId) {
+              videoDetailsDAO.removeById(id)
+              item.fileMetaId.map(binaryStorageService.remove(_, identity))
+              ctx.complete(item)
+            } else {
+              ctx.reject(AuthorizationFailedRejection)
+            }
+          }
+          case None => ctx.complete(StatusCodes.NotFound)
+        }
       }
     }
   }
